Match the number of interrupts to the number of queues
The problem was that the length of `interrupts` was always 1 while it's
expected to be equal to the length of `queues` and `queue_evts`.
Since `interrupts` was shorter, some queues were uninitialized as we are
calling `Iterator::zip` for `queues` and `interrupts`.

This was introduced at [1]. So, the async block device has been broken
since that commit.

[1]
https://github.com/dgreid/crosvm/commit/e8325a5927c039809596dfd1155fe0263ed2f637

Change-Id: Ic6d8ec9353f0251bf1829dd9c941540006ff77e6